[refactor]: mypage에서 profileView 재사용을 위한 컴포넌트화 (#76)
